face-recognition-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymilvus==2.3.0
marshmallow==3.19.0
redis==5.0.1
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.1
onnxruntime==1.16.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
httpx==0.25.0
aiofiles==23.2.1